+++ kite/src/main/kotlin/com/tang/kite/sql/provider/ProviderType.kt
@@ -6,8 +6,11 @@ package com.tang.kite.sql.provider
  * @author Tang
  */
 enum class ProviderType {
+
     POSTGRESQL,
     MYSQL,
+    SQLITE,
     DERBY,
     H2
+
 }